+++ CHANGELOG.md
@@ -5,9 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-03-29
+
+### Added
+* We've added the ability to override settings in the default config file by passing in a config file with the settings needing to be overridden when instantiating a Unity object. [56](https://github.com/unity-sds/unity-py/issues/56)
+* Collection creation (create_collection) through dataService library
+* Added support for defining custom metadata for project and venue.
+* Added methods to return STAC content instead of unity domain objects if requested
+* Added properties parsing of stac metadata to dataset objects
+### Fixed
+### Changed
+* Updated get_collections and get_collection_data to support limit parameter.
+* Updated get_collection_data to support filter parameter which takes CQL string.
+### Removed
+### Security
+### Deprecated
 
 
-## Unreleased [0.3.0]
+## [0.3.0] - 2024-02-12
+
 ### Added
 ### Fixed
 * fixed an issue with encoding a json deploy request twice [71](https://github.com/unity-sds/unity-py/issues/71)